requirements.txt
streamlit
streamlit-webrtc
mediapipe
opencv-python
torch
torchvision
numpy
Pillow
av
